+++ src/components/header/Header.js
@@ -1,12 +1,6 @@
-import React, { useEffect, useState } from "react";
-import { getfoldersData } from "../../api/folder";
+import React from "react";
 import Folder from "../../folder/Folder";
-import styles from "./Header.module.css";
-import Profile from "../profile/Profile";
-import ProfileSkeleton from "../skeleton/ItemSkeleton";
-import useFetch from "../../hooks/useFetch";
 import FolderSkeleton from "../skeleton/FolderSkeleton";
-
 export default function Header({ data, isLoading }) {
   return (
     <div>
