--- src/components/header/Header.module.css
@@ -1,26 +1,0 @@
-.container {
-  border: 5px solid red;
-  background-color: #edf7ff;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  row-gap: 2.5rem;
-  padding: 3rem;
-  /* tablet */
-  @media (min-width: 768px) {
-    padding: 5rem;
-  }
-  /* descktop*/
-  @media (min-width: 1200px) {
-    padding: 5rem;
-  }
-}
-
-.text__codeit {
-  font-size: 2rem;
-  font-weight: bold;
-}
-.text__favorite {
-  font-size: 3rem;
-  font-weight: bold;
-}
