+++ src/components/nav/Nav.js
@@ -1,38 +1,24 @@
 import React, { useEffect, useState } from "react";
-
 import { getUsersData } from "../../api/users";
 import styled from "styled-components";
 import { ReactComponent as Logo } from "../../assets/images/logo.svg";
 import { ReactComponent as Person } from "../../assets/images/person.svg";
 import { ReactComponent as Login } from "../../assets/images/login.svg";
 import Profile from "../profile/Profile";
-
-const ResponSiveNavbar = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 3.2rem;
-  background-color: #edf7ff;
-  height: 7rem;
-  /* tablet */
-  @media (min-width: 768px) {
-    height: 9.5rem;
-  }
-  /* descktop*/
-  @media (min-width: 1200px) {
-    padding: 0 20rem;
-    height: 9.5rem;
-  }
-`;
+import styles from "./Nav.module.css";
+import useFetch from "../../hooks/useFetch";
 
 const ResponSiveProfie = styled.div`
   display: flex;
   align-items: center;
 `;
+
 export default function Nav() {
   const [isUser, setIsUser] = useState(false);
   const [userEmail, setUserEmail] = useState("");
+  // const response = useFetch("https://bootcamp-api.codeit.kr/api/sample/user");
+  // const [data, isLoading] = response;
+
   const handlUserDate = async () => {
     const { email } = await getUsersData();
     if (email) {
@@ -47,13 +33,14 @@ export default function Nav() {
   }, []);
   return (
     <div>
-      <ResponSiveNavbar>
+      <div className={styles.container}>
         <Logo />
+
         <ResponSiveProfie>
           {isUser ? <Person /> : <Login />}
           {isUser && <span className="user-email">{userEmail}</span>}
         </ResponSiveProfie>
-      </ResponSiveNavbar>
+      </div>
     </div>
   );
 }
