+++ src/components/nav/Nav.module.css
@@ -16,15 +16,3 @@
     height: 9.5rem;
   }
 }
-
-.user-email {
-  display: none;
-
-  @media (min-width: 768px) {
-    display: inline-block;
-  }
-  /* descktop*/
-  @media (min-width: 1200px) {
-    display: inline-block;
-  }
-}
