+++ src/components/profile/Profile.js
@@ -1,20 +1,5 @@
-import React, { useEffect, useState } from "react";
-import useFetch from "../../hooks/useFetch";
-import styles from "./Profile.module.css";
-export default function Profile({ data, isLoading }) {
-  // const [personName, setPersonName] = useState("");
-  // const [personImage, setPersonImage] = useState("");
-  // return (
-  //   // <div>
-  //   //   {!isLoading ? (
-  //   //     <>
-  //   //       {" "}
-  //   //       <img src={profieImageSource} alt={name} width="200"></img>
-  //   //       <p className={styles.text__codeit}>{name}</p>
-  //   //     </>
-  //   //   ) : (
-  //   //     <p>로딩중</p>
-  //   //   )}
-  //   // </div>
-  // );
+import React from "react";
+
+export default function Profile() {
+  return <div></div>;
 }
