+++ src/pages/FolderPage.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function FolderPage() {
+  return <div>dddd</div>;
+}
